requirements.txt
# requirements for the python scripts
numpy >= 1.20.1
Pillow >= 8.0.1